Internal Zip class now uses a filtering callback to determine files to exclude.
Bug 27840 asks for richer logic to determine which files to exclude from a packed extension. This change sets up the internals to add that logic, but it doesn't yet actually add it; that's why there are no corresponding unit-test changes. That's because this is my first Chrome CL, and I wanted to get something minimal through the commit/build process.

TEST=none (no external functionality should have changed)
BUG=27840

Review URL: http://codereview.chromium.org/7677024

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@97837 0039d316-1c4b-4281-b951-d872f2087c98