Add local histogram to see what UMA pageloads are
This adds a histogram that gets a sample each time the
UMA pageload counter is incremented.  The result is that
any time you refresh:
about:histograms/Chrome.UmaPageloadCounter
You'll see how many pageloads have taken place during
the session.  Be warned: Each refresh of that histogram
page will also add 1 to the tally ;-).

If you want to use this patch in a personal build, be
sure to modify browser_main.cc so that metrics->Start()
is called (even if this is not an official build).
Review URL: http://codereview.chromium.org/7792070

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@99289 0039d316-1c4b-4281-b951-d872f2087c98