clang update script: Remove --lldb, add --mac-only, --force-local-build
lldb doesn't build on linux, and is difficult to build on mac and ships
with xcode 4 anyway (_and_ isn't all that good), so remove the option for
it for now.

Add a --mac-only option that makes the script exit early on non-mac systems
(to be used to add a gclient hook that pulls clang for all devs on mac).

Add a --force-local-build option that can be used to hack on clang.

BUG=none
TEST=none

Review URL: http://codereview.chromium.org/7826023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@99419 0039d316-1c4b-4281-b951-d872f2087c98