Fix URLFetcherTest.CancelAll.
I don't know what I was smoking before when I wrote this test. It was obviously racy.
Fix it by moving the EXPECTs and CancelAll call to the IO thread.
The last EXPECT is fine on the main test thread since it should be guaranteed to be 0.

BUG=89863,88221
TEST=manual

Review URL: http://codereview.chromium.org/7881001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@100861 0039d316-1c4b-4281-b951-d872f2087c98