Fix crash. Call GetResourceContext() on IO thread.
When we cancel the request at PluginService, we the PluginProcessHost::Client will go away. Therefore it's not safe to call it on another thread, we get a race condition that way. It should probably only be used on one thread anyway. Call GetResourceContext() on IO thread and pass it through to the FILE thread.

BUG=96759
TEST=none

Review URL: http://codereview.chromium.org/7908015

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@101427 0039d316-1c4b-4281-b951-d872f2087c98