Fix the Tangent build.
A recent refactoring caused the Tangent build to crash on start-up.
This is a temporary work-around that will avoid the crash until the IME
team in Tokyo can fix it.

This is the CL that introduced the crash:
http://codereview.chromium.org/7887032/

I suspect that GetSupportedInputMethods needs to be calling
CreateInputMethodDescriptor, but I could not see a good way to do that
from a static context because of the lack of an available Whitelist.

BUG=none
TEST=compile & run touchui build, it doesn't crash anymore

Review URL: http://codereview.chromium.org/7918017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@101489 0039d316-1c4b-4281-b951-d872f2087c98