Experimentally strictify third-party cookie blocking.
This CL turns on strict third-party cookie blocking by forcing the inclusion of
the `kBlockReadingThirdPartyCookies` command-line flag. The goal is to determine
the effect of a change in cookie-blocking behavior; if we Break The Web™ for
Canary users, we'll drop it. If not, we'll re-evaluate the default behavior of
the `kBlockThirdPartyCookies` pref.

I intend to revert this change after a few Canary cycles.

BUG=98241
TEST=Start Chrome. Is the command-line flag set?


Review URL: http://codereview.chromium.org/8060030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103112 0039d316-1c4b-4281-b951-d872f2087c98