Don't show hidden devices in UI (exept from imageburner).
At the moment CrosDisks does not send signals for devices that should not be automounted.
Unfortunately, this breaks imageburner, so we are changing CrosDisks to start sending events for those devices, but marks them hidden.
This CL should ensure that those devices get processed only in context of imageburner.

libcros change:
http://gerrit.chromium.org/gerrit/#change,8216

TEST=Verified that all attached removable devices show up on   chrome://imageburner
Checked that hidden devices are not shown in file browser, or attempted to be mounted.
BUG=chromium-os:20476

Review URL: http://codereview.chromium.org/8049002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@103965 0039d316-1c4b-4281-b951-d872f2087c98