profile: Add a way for each ProfileKeyedServiceFactory to specify its user prefernces.
To allow us to compile individual features in/out of a chrome build, we need to
attack browser_prefs.cc, which is a giant static registration. This used to be
done through another big code path starting in ProfileImpl which call
browser::RegisterUserPrefs() which linked to other static methods. Now this is
done through overridding ProfileKeyedServiceFactory::RegisterUserPrefs(), and
there are no additional dependencies.

BUG=none
TEST=none


Review URL: http://codereview.chromium.org/7901027

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@104412 0039d316-1c4b-4281-b951-d872f2087c98