chromeos: Load machine statistics in the FILE thread.
The loading starts soon after FILE thread is created. If the statistics
are requested before loading is done, the requests are blocked until
the loading is done.

Hopefully, statistics are loaded before requested as of now
for regular sessions. See comments in statistics_provider.cc for details.

Ideally we don't want to block at all, but there are many clients of 
StatisticsProvider and changing them to use a callback is not feasible.


The change depends on http://codereview.chromium.org/8200009/

BUG=chromium-os:5603
TEST=confirm that statistics are loaded properly on chrome os (add --vmodule=name_value_pairs_parser=1,statistics_provider=1

patch from issue 8200009


Review URL: http://codereview.chromium.org/8202017

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@104765 0039d316-1c4b-4281-b951-d872f2087c98