WriteInto handles length_with_null values of 1 better.
In debug builds on VS2010, the old code triggered a debug assertion in
xstring, since [0] was being accessed and the size of the array was 0.

A |length_with_null| of 1 is considered a valid use of this function, so
an alternate implementation is used which returns the data() pointer.

BUG=None
TEST=net_unittests --gtest_filter=*Digest* on debug on VS2010 works.


Review URL: http://codereview.chromium.org/8143001

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106062 0039d316-1c4b-4281-b951-d872f2087c98