Sync Promo: Replace original startup page with sync promo
Currently the sync promo tab is opened in addition to the startup tabs that a user would normally see.

After some discussion we've agreed that this looks to cluttered.

This CL changes the startup so that the sync promo tab replaces the first tab. Once the user completes or skips the promo we navigate to the original startup tab.

BUG=99743
TEST=


Review URL: http://codereview.chromium.org/8352022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@106419 0039d316-1c4b-4281-b951-d872f2087c98