IRT name was incorrect Activate IRT building on ARM for NaCl.
In order for this to be on for real, the PNaCl toolchain still needs to
available

Since NaCl is still turned off by default on ARM, this code should be a no-op
for all other cases.

This also moves the compilation a step forward for ARM. 

There is a bootstrap dependency on memset on ARM - so SYSROOT on the
ChromiumOS chroot is required to gain access to libc.

R=bradnelson,robertm
BUG=http://code.google.com/p/chromium/issues/detail?id=61695
BUG=http://code.google.com/p/chromium/issues/detail?id=38909
BUG=http://code.google.com/p/nativeclient/issues/detail?id=135


Review URL: http://codereview.chromium.org/8375035

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@107040 0039d316-1c4b-4281-b951-d872f2087c98