chromeos: Simplify power supply info in PowerLibrary
Factoring out the status_ field of PowerLibrary, along with the numerous
accessor functions.  Instead, the power status will be used as a local
variable, passed along to observers as an argument.

Also, created a local version of the power supply status struct.  This is the
struct that will be used when the libcros version of the struct is removed.

See satorux's comments here:
http://codereview.chromium.org/8271024/

BUG=chromium-os:16558
TEST=Make sure power button icon behaves properly, on both AC and battery.

Signed-off-by: Simon Que <sque@chromium.org>

R=satorux@chromium.org,stevenjb@chromium.org


Review URL: http://codereview.chromium.org/8347016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@107465 0039d316-1c4b-4281-b951-d872f2087c98