Added logic that will retry to reload+reconnect the mobile payment portal page 5 times before it eventually gives up and shows service unavailable error.
Expanded frame sniffer to send an event when a frame which is being watched gets finally loaded. From the sent error (or its absence) and this new event, we can finally reconstruct on the browser side if a given iframe was loaded successfully or not.  

BUG=chromium-os:20525, chromium-os:21738
TEST=tricky one to test since it repros on 3g network flakyness. one can edit /usr/lib/cromo/config/madison.ini and put a bad value for olp_root there to trigger this.
Review URL: http://codereview.chromium.org/8341037

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@107613 0039d316-1c4b-4281-b951-d872f2087c98